start on scraper

+++ models/price.go
@@ -11,7 +11,7 @@ import (
 type Price struct {
     Id *int64 `bun:"id,pk,autoincrement"`
     MerchantId *int64
-    Value *int64 `bun:",notnull"`
+    Value *float64 `bun:",notnull"`
     CreatedAt time.Time `bun:",nullzero,notnull,default:current_timestamp"`
     UpdatedAt time.Time `bun:",nullzero,notnull,default:current_timestamp"`
 }
